backend-tourapp/requirements.txt
Django==4.2.10
djangorestframework==3.14.0
django-cors-headers==4.3.1
dj-rest-auth==5.0.1
django-filter==23.5
Pillow==10.1.0
python-dotenv==1.0.0
django-environ==0.11.2
drf-yasg==1.21.7
django-jazzmin==2.6.0
